Refactor typewriter text component
- Move typewriter text logic to a new component `TypeWriterInitialText`.
- Clean up `HomePage` by importing and using the new component.
- Ensure visual consistency and reuse code effectively.

(✿◠‿◠)
